[*] Refactory na classe CertificadoDigita metodos de assinatura de Xml movido para a classe Sha1XmlSigning. [*] Refactory na classe CertificadoDigital metodos de validação de xml movido para classe XmlValidation. [*] Ajustes nas classes DFeServiceClientBase e DFeSoap12ServiceClient para ficar compativel com .Net Standard 2.0.
