+++ Nuget/ACBr.Net.DFe.Core.nuspec
@@ -29,5 +29,6 @@
 		<file src="..\bin\Release\Net40\ACBr.Net.DFe.Core.XML" target="lib\net40\ACBr.Net.DFe.Core.XML" />
         <file src="..\bin\Release\netstandard2.0\ACBr.Net.DFe.Core.dll" target="lib\netstandard2.0\ACBr.Net.DFe.Core.dll" />
         <file src="..\bin\Release\netstandard2.0\ACBr.Net.DFe.Core.xml" target="lib\netstandard2.0\ACBr.Net.DFe.Core.xml" />
+		<file src="..\bin\Release\netstandard2.0\ACBr.Net.DFe.Core.deps.json" target="lib\netstandard2.0\ACBr.Net.DFe.Core.deps.json" />		
 	</files>
 </package>
