+++ src/ACBr.Net.DFe.Core.Shared/CertificadoDigital.cs
@@ -32,214 +32,101 @@
 using ACBr.Net.Core.Exceptions;
 using ACBr.Net.Core.Extensions;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Security.Cryptography.X509Certificates;
-using System.Security.Cryptography.Xml;
-using System.Xml;
-using System.Xml.Schema;
 
 namespace ACBr.Net.DFe.Core
 {
-	/// <summary>
-	/// Classe CertificadoDigital.
-	/// </summary>
-	public static class CertificadoDigital
-	{
-		#region Methods
+    /// <summary>
+    /// Classe CertificadoDigital.
+    /// </summary>
+    public static class CertificadoDigital
+    {
+        #region Methods
 
-		/// <summary>
-		/// Assina a XML usando o certificado informado.
-		/// </summary>
-		/// <param name="xml">O Xml.</param>
-		/// <param name="pUri">A Uri da referencia.</param>
-		/// <param name="pNode">O node onde sera feito o append da assinatura.</param>
-		/// <param name="pCertificado">O certificado.</param>
-		/// <param name="comments">Se for <c>true</c> vai inserir a tag #withcomments no transform.</param>
-		/// <returns>System.String.</returns>
-		/// <exception cref="ACBrDFeException">Erro ao efetuar assinatura digital.</exception>
-		/// <exception cref="ACBrDFeException">Erro ao efetuar assinatura digital.</exception>
-		public static string AssinarXml(string xml, string pUri, string pNode, X509Certificate2 pCertificado, bool comments = false)
-		{
-			try
-			{
-				var doc = new XmlDocument();
-				doc.LoadXml(xml);
-				AssinarDocumento(doc, pUri, pNode, pCertificado, comments);
-				return doc.AsString();
-			}
-			catch (Exception ex)
-			{
-				throw new ACBrDFeException("Erro ao efetuar assinatura digital.", ex);
-			}
-		}
+        /// <summary>
+        /// Busca certificados instalado se informado uma serie
+        /// senão abre caixa de dialogos de certificados.
+        /// </summary>
+        /// <param name="cerSerie">Serie do certificado.</param>
+        /// <returns>X509Certificate2.</returns>
+        /// <exception cref="System.Exception">
+        /// Nenhum certificado digital foi selecionado ou o certificado selecionado está com problemas.
+        /// or
+        /// Certificado digital não encontrado
+        /// or
+        /// </exception>
+        public static X509Certificate2 SelecionarCertificado(string cerSerie = "")
+        {
+            var store = new X509Store("MY", StoreLocation.CurrentUser);
+            store.Open(OpenFlags.MaxAllowed | OpenFlags.ReadOnly);
 
-		/// <summary>
-		/// Assina Multiplos elementos dentro da Xml.
-		/// </summary>
-		/// <param name="xml">O Xml.</param>
-		/// <param name="pSignNode">O node inicial para o calculo da assinatura.</param>
-		/// <param name="pUri">A Uri da referencia.</param>
-		/// <param name="pNode">O node onde sera feito o append da assinatura</param>
-		/// <param name="pCertificado">O certificado.</param>
-		/// <param name="comments">Se for <c>true</c> vai inserir a tag #withcomments no transform.</param>
-		/// <returns>System.String.</returns>
-		/// <exception cref="ACBrDFeException">Erro ao efetuar assinatura digital.</exception>
-		/// <exception cref="ACBrDFeException">Erro ao efetuar assinatura digital.</exception>
-		public static string AssinarXmlTodos(string xml, string pSignNode, string pUri, string pNode, X509Certificate2 pCertificado, bool comments = false)
-		{
-			try
-			{
-				var doc = new XmlDocument();
-				doc.LoadXml(xml);
+            try
+            {
+                var certificates = store.Certificates.Find(X509FindType.FindByTimeValid, DateTime.Now, true)
+                    .Find(X509FindType.FindByKeyUsage, X509KeyUsageFlags.DigitalSignature, false);
 
-				var xmlElements = doc.GetElementsByTagName(pSignNode).Cast<XmlElement>().ToArray();
-				Guard.Against<ACBrDFeException>(!xmlElements.Any(), "Nome do elemento de assinatura incorreto");
+                X509Certificate2Collection certificadosSelecionados;
 
-				foreach (var element in xmlElements)
-				{
-					var docElement = new XmlDocument();
-					docElement.LoadXml(element.OuterXml);
-					AssinarDocumento(docElement, pUri, pNode, pCertificado, comments);
-
-					// ReSharper disable once AssignNullToNotNullAttribute
-					var signedElement = doc.ImportNode(docElement.DocumentElement, true);
-					element.ParentNode?.ReplaceChild(signedElement, element);
-				}
-
-				return doc.AsString();
-			}
-			catch (Exception ex)
-			{
-				throw new ACBrDFeException("Erro ao efetuar assinatura digital.", ex);
-			}
-		}
-
-		private static void AssinarDocumento(XmlDocument doc, string pUri, string pNode, X509Certificate2 pCertificado, bool comments = false)
-		{
-			Guard.Against<ArgumentException>(!pUri.IsEmpty() && doc.GetElementsByTagName(pUri).Count != 1, "Referencia invalida ou não é unica.");
-
-			//Adiciona Certificado ao Key Info
-			var keyInfo = new KeyInfo();
-			keyInfo.AddClause(new KeyInfoX509Data(pCertificado));
-
-			//Seta chaves
-			var signedDocument = new SignedXml(doc)
-			{
-				SigningKey = pCertificado.PrivateKey,
-				KeyInfo = keyInfo
-			};
-
-			var uri = pUri.IsEmpty() ? "" : $"#{doc.GetElementsByTagName(pUri)[0].Attributes?["Id"]?.InnerText}";
-
-			// Cria referencia
-			var reference = new Reference { Uri = uri };
-
-			// Adiciona transformação a referencia
-			reference.AddTransform(new XmlDsigEnvelopedSignatureTransform());
-			reference.AddTransform(new XmlDsigC14NTransform(comments));
-
-			// Adiciona referencia ao xml
-			signedDocument.AddReference(reference);
-
-			// Calcula Assinatura
-			signedDocument.ComputeSignature();
-
-			// Pega representação da assinatura
-			var xmlDigitalSignature = signedDocument.GetXml();
-
-			// Adiciona ao doc XML
-			var xmlElement = doc.GetElementsByTagName(pNode).Cast<XmlElement>().FirstOrDefault();
-			Guard.Against<ACBrDFeException>(xmlElement == null, "Nome do elemento para o append da assinatura incorreto");
-
-			var element = doc.ImportNode(xmlDigitalSignature, true);
-			xmlElement.AppendChild(element);
-		}
-
-		/// <summary>
-		/// Busca certificados instalado se informado uma serie
-		/// senão abre caixa de dialogos de certificados.
-		/// </summary>
-		/// <param name="cerSerie">Serie do certificado.</param>
-		/// <returns>X509Certificate2.</returns>
-		/// <exception cref="System.Exception">
-		/// Nenhum certificado digital foi selecionado ou o certificado selecionado está com problemas.
-		/// or
-		/// Certificado digital não encontrado
-		/// or
-		/// </exception>
-		public static X509Certificate2 SelecionarCertificado(string cerSerie = "")
-		{
-			var store = new X509Store("MY", StoreLocation.CurrentUser);
-			store.Open(OpenFlags.MaxAllowed | OpenFlags.ReadOnly);
-
-			try
-			{
-				var certificates = store.Certificates.Find(X509FindType.FindByTimeValid, DateTime.Now, true)
-					.Find(X509FindType.FindByKeyUsage, X509KeyUsageFlags.DigitalSignature, false);
-
-				X509Certificate2Collection certificadosSelecionados;
-
-				if (cerSerie.IsEmpty())
-				{
+                if (cerSerie.IsEmpty())
+                {
 #if NETSTANDARD2_0
-					throw new ACBrDFeException("Numero de seri obrigatorio.");
+                    throw new ACBrDFeException("Numero de seri obrigatorio.");
 #else
 					certificadosSelecionados = X509Certificate2UI.SelectFromCollection(certificates, "Certificados Digitais",
                         "Selecione o Certificado Digital para uso no aplicativo", X509SelectionFlag.SingleSelection);
 #endif
-				}
-				else
-				{
-					certificadosSelecionados = certificates.Find(X509FindType.FindBySerialNumber, cerSerie, false);
-					Guard.Against<ACBrDFeException>(certificadosSelecionados.Count == 0, "Certificado digital não encontrado");
-				}
+                }
+                else
+                {
+                    certificadosSelecionados = certificates.Find(X509FindType.FindBySerialNumber, cerSerie, false);
+                    Guard.Against<ACBrDFeException>(certificadosSelecionados.Count == 0, "Certificado digital não encontrado");
+                }
 
-				var certificado = certificadosSelecionados.Count < 1 ? null : certificadosSelecionados[0];
-				return certificado;
-			}
-			catch (Exception ex)
-			{
-				throw new ACBrDFeException("Erro ao selecionar o certificado", ex);
-			}
-			finally
-			{
-				store.Close();
-			}
-		}
+                var certificado = certificadosSelecionados.Count < 1 ? null : certificadosSelecionados[0];
+                return certificado;
+            }
+            catch (Exception ex)
+            {
+                throw new ACBrDFeException("Erro ao selecionar o certificado", ex);
+            }
+            finally
+            {
+                store.Close();
+            }
+        }
 
-		/// <summary>
-		/// Seleciona um certificado informando o caminho e a senha.
-		/// </summary>
-		/// <param name="caminho">O caminho.</param>
-		/// <param name="senha">A senha.</param>
-		/// <returns>X509Certificate2.</returns>
-		/// <exception cref="System.Exception">Arquivo do Certificado digital não encontrado</exception>
-		public static X509Certificate2 SelecionarCertificado(string caminho, string senha)
-		{
-			Guard.Against<ArgumentNullException>(caminho.IsEmpty(), "Caminho do arquivo não poder ser nulo ou vazio !");
-			Guard.Against<ArgumentException>(!File.Exists(caminho), "Arquivo do Certificado digital não encontrado !");
+        /// <summary>
+        /// Seleciona um certificado informando o caminho e a senha.
+        /// </summary>
+        /// <param name="caminho">O caminho.</param>
+        /// <param name="senha">A senha.</param>
+        /// <returns>X509Certificate2.</returns>
+        /// <exception cref="System.Exception">Arquivo do Certificado digital não encontrado</exception>
+        public static X509Certificate2 SelecionarCertificado(string caminho, string senha)
+        {
+            Guard.Against<ArgumentNullException>(caminho.IsEmpty(), "Caminho do arquivo não poder ser nulo ou vazio !");
+            Guard.Against<ArgumentException>(!File.Exists(caminho), "Arquivo do Certificado digital não encontrado !");
 
-			var cert = new X509Certificate2(caminho, senha);
-			return cert;
-		}
+            var cert = new X509Certificate2(caminho, senha);
+            return cert;
+        }
 
-		/// <summary>
-		/// Seleciona um certificado passando um array de bytes.
-		/// </summary>
-		/// <param name="certificado">O certificado.</param>
-		/// <param name="senha">A senha.</param>
-		/// <returns>X509Certificate2.</returns>
-		/// <exception cref="System.Exception">Arquivo do Certificado digital não encontrado</exception>
-		public static X509Certificate2 SelecionarCertificado(byte[] certificado, string senha = "")
-		{
-			Guard.Against<ArgumentNullException>(certificado != null, "O certificado não poder ser nulo ou vazio !");
-			Guard.Against<ArgumentException>(certificado.Length > 0, "O tamanhado do certificado não pode ser zero !");
+        /// <summary>
+        /// Seleciona um certificado passando um array de bytes.
+        /// </summary>
+        /// <param name="certificado">O certificado.</param>
+        /// <param name="senha">A senha.</param>
+        /// <returns>X509Certificate2.</returns>
+        /// <exception cref="System.Exception">Arquivo do Certificado digital não encontrado</exception>
+        public static X509Certificate2 SelecionarCertificado(byte[] certificado, string senha = "")
+        {
+            Guard.Against<ArgumentNullException>(certificado != null, "O certificado não poder ser nulo ou vazio !");
+            Guard.Against<ArgumentException>(certificado.Length > 0, "O tamanhado do certificado não pode ser zero !");
 
-			var cert = new X509Certificate2(certificado, senha);
-			return cert;
-		}
+            var cert = new X509Certificate2(certificado, senha);
+            return cert;
+        }
 
 #if !NETSTANDARD2_0
 
@@ -256,85 +143,6 @@ namespace ACBr.Net.DFe.Core
 
 #endif
 
-		/// <summary>
-		/// Valida o arquivo XML com o schema informado.
-		/// </summary>
-		/// <param name="arquivoXml">The arquivo XML.</param>
-		/// <param name="schema">The schema nf.</param>
-		/// <param name="erros">The erro.</param>
-		/// <param name="avisos">The avisos.</param>
-		/// <returns><c>true</c> if XXXX, <c>false</c> otherwise.</returns>
-		public static bool ValidarXml(string arquivoXml, string schema, out string[] erros, out string[] avisos)
-		{
-			var errorList = new List<string>();
-			var avisosList = new List<string>();
-
-			if (string.IsNullOrEmpty(arquivoXml))
-			{
-				errorList.Add("Arquivo da nota fiscal não encontrado.");
-				erros = errorList.ToArray();
-				avisos = avisosList.ToArray();
-				return false;
-			}
-
-			if (!File.Exists(schema))
-			{
-				errorList.Add("Arquivo de Schema não encontrado.");
-				erros = errorList.ToArray();
-				avisos = avisosList.ToArray();
-				return false;
-			}
-
-			try
-			{
-				var xmlSchema = XmlSchema.Read(new XmlTextReader(schema), (sender, args) =>
-				{
-					switch (args.Severity)
-					{
-						case XmlSeverityType.Warning:
-							// ReSharper disable once AccessToModifiedClosure
-							avisosList.Add(args.Message);
-							break;
-
-						case XmlSeverityType.Error:
-							// ReSharper disable once AccessToModifiedClosure
-							errorList.Add(args.Message);
-							break;
-					}
-
-					// Erro na validação do schema XSD
-					if (args.Exception != null)
-					{
-						// ReSharper disable once AccessToModifiedClosure
-						errorList.Add("\nErro: " + args.Exception.Message + "\nLinha " + args.Exception.LinePosition + " - Coluna "
-								  + args.Exception.LineNumber + "\nSource: " + args.Exception.SourceUri);
-					}
-				});
-
-				var settings = new XmlReaderSettings();
-				settings.ValidationType = ValidationType.Schema;
-				settings.Schemas.Add(xmlSchema);
-
-				using (var xmlReader = XmlReader.Create(new StringReader(arquivoXml), settings))
-				{
-					while (xmlReader.Read())
-					{
-					}
-				}
-			}
-			catch (Exception exception)
-			{
-				errorList.Add(exception.Message);
-			}
-
-			erros = errorList.ToArray();
-			avisos = avisosList.ToArray();
-			errorList = null;
-			avisosList = null;
-
-			return (erros.Length < 1);
-		}
-
-		#endregion Methods
-	}
+        #endregion Methods
+    }
 }
