+++ src/ACBr.Net.DFe.Core.Shared/Common/DFeDocument.cs
@@ -51,15 +51,13 @@ namespace ACBr.Net.DFe.Core.Common
 		/// <returns>TDocument.</returns>
 		public static TDocument Load(string document, Encoding encoding = null)
 		{
-			using (var serializer = new DFeSerializer(typeof(TDocument)))
+			var serializer = new DFeSerializer(typeof(TDocument));
+			if (encoding != null)
 			{
-				if (encoding != null)
-				{
-					serializer.Options.Encoder = encoding;
-				}
-
-				return (TDocument)serializer.Deserialize(document);
+				serializer.Options.Encoder = encoding;
 			}
+
+			return (TDocument)serializer.Deserialize(document);
 		}
 
 		/// <summary>
@@ -70,15 +68,13 @@ namespace ACBr.Net.DFe.Core.Common
 		/// <returns>TDocument.</returns>
 		public static TDocument Load(Stream document, Encoding encoding = null)
 		{
-			using (var serializer = new DFeSerializer(typeof(TDocument)))
+			var serializer = new DFeSerializer(typeof(TDocument));
+			if (encoding != null)
 			{
-				if (encoding != null)
-				{
-					serializer.Options.Encoder = encoding;
-				}
-
-				return (TDocument)serializer.Deserialize(document);
+				serializer.Options.Encoder = encoding;
 			}
+
+			return (TDocument)serializer.Deserialize(document);
 		}
 
 		/// <summary>
@@ -109,22 +105,21 @@ namespace ACBr.Net.DFe.Core.Common
 		/// <returns>TDocument.</returns>
 		public void Save(string path, DFeSaveOptions options = DFeSaveOptions.None, Encoding encoding = null)
 		{
-			using (var serializer = new DFeSerializer(typeof(TDocument)))
+			var serializer = new DFeSerializer(typeof(TDocument));
+
+			if (!options.HasFlag(DFeSaveOptions.None))
 			{
-				if (!options.HasFlag(DFeSaveOptions.None))
-				{
-					serializer.Options.RemoverAcentos = options.HasFlag(DFeSaveOptions.RemoveAccents);
-					serializer.Options.FormatarXml = !options.HasFlag(DFeSaveOptions.DisableFormatting);
-					serializer.Options.OmitirDeclaracao = !options.HasFlag(DFeSaveOptions.OmitDeclaration);
-				}
-
-				if (encoding != null)
-				{
-					serializer.Options.Encoder = encoding;
-				}
-
-				serializer.Serialize(this, path);
+				serializer.Options.RemoverAcentos = options.HasFlag(DFeSaveOptions.RemoveAccents);
+				serializer.Options.FormatarXml = !options.HasFlag(DFeSaveOptions.DisableFormatting);
+				serializer.Options.OmitirDeclaracao = !options.HasFlag(DFeSaveOptions.OmitDeclaration);
 			}
+
+			if (encoding != null)
+			{
+				serializer.Options.Encoder = encoding;
+			}
+
+			serializer.Serialize(this, path);
 		}
 
 		/// <summary>
@@ -136,22 +131,21 @@ namespace ACBr.Net.DFe.Core.Common
 		/// <returns>TDocument.</returns>
 		public void Save(Stream stream, DFeSaveOptions options = DFeSaveOptions.None, Encoding encoding = null)
 		{
-			using (var serializer = new DFeSerializer(typeof(TDocument)))
+			var serializer = new DFeSerializer(typeof(TDocument));
+
+			if (!options.HasFlag(DFeSaveOptions.None))
 			{
-				if (!options.HasFlag(DFeSaveOptions.None))
-				{
-					serializer.Options.RemoverAcentos = options.HasFlag(DFeSaveOptions.RemoveAccents);
-					serializer.Options.FormatarXml = !options.HasFlag(DFeSaveOptions.DisableFormatting);
-					serializer.Options.OmitirDeclaracao = !options.HasFlag(DFeSaveOptions.OmitDeclaration);
-				}
-
-				if (encoding != null)
-				{
-					serializer.Options.Encoder = encoding;
-				}
-
-				serializer.Serialize(this, stream);
+				serializer.Options.RemoverAcentos = options.HasFlag(DFeSaveOptions.RemoveAccents);
+				serializer.Options.FormatarXml = !options.HasFlag(DFeSaveOptions.DisableFormatting);
+				serializer.Options.OmitirDeclaracao = !options.HasFlag(DFeSaveOptions.OmitDeclaration);
 			}
+
+			if (encoding != null)
+			{
+				serializer.Options.Encoder = encoding;
+			}
+
+			serializer.Serialize(this, stream);
 		}
 	}
 }
