+++ src/ACBr.Net.DFe.Core.Shared/Serializer/SerializerOptions.cs
@@ -30,75 +30,66 @@
 // ***********************************************************************
 
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Security.Cryptography.X509Certificates;
 using System.Text;
 
 namespace ACBr.Net.DFe.Core.Serializer
 {
-    /// <summary>
-    /// Class SerializerOptions. This class cannot be inherited.
-    /// </summary>
-    public class SerializerOptions
-    {
-        #region Constructors
+	/// <summary>
+	/// Class SerializerOptions. This class cannot be inherited.
+	/// </summary>
+	public class SerializerOptions
+	{
+		#region Constructors
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DFeSerializer" /> class.
-        /// </summary>
-        internal SerializerOptions()
-        {
-            ErrosAlertas = new List<string>();
-            FormatoAlerta = "TAG:%TAG% ID:%ID%/%TAG%(%DESCRICAO%) - %MSG%.";
-            RemoverAcentos = false;
-            FormatarXml = true;
-            AssinarXML = false;
-            Encoder = Encoding.UTF8;
-        }
+		/// <summary>
+		/// Initializes a new instance of the <see cref="DFeSerializer" /> class.
+		/// </summary>
+		internal SerializerOptions()
+		{
+			ErrosAlertas = new List<string>();
+			FormatoAlerta = "TAG:%TAG% ID:%ID%/%TAG%(%DESCRICAO%) - %MSG%.";
+			RemoverAcentos = false;
+			FormatarXml = true;
+			Encoder = Encoding.UTF8;
+		}
 
-        #endregion Constructors
+		#endregion Constructors
 
-        #region Propriedades
+		#region Propriedades
 
-        public bool RemoverAcentos { get; set; }
+		public bool RemoverAcentos { get; set; }
 
-        public bool FormatarXml { get; set; }
+		public bool FormatarXml { get; set; }
 
-        public bool OmitirDeclaracao { get; set; }
+		public bool OmitirDeclaracao { get; set; }
 
-        public bool AssinarXML { get; set; }
+		public Encoding Encoder { get; set; }
 
-        public string SignUri { get; set; }
+		public List<string> ErrosAlertas { get; }
 
-        public X509Certificate Certificado { get; set; }
+		public string FormatoAlerta { get; set; }
 
-        public Encoding Encoder { get; set; }
+		#endregion Propriedades
 
-        public List<string> ErrosAlertas { get; }
+		#region Methods
 
-        public string FormatoAlerta { get; set; }
+		internal void AddAlerta(string id, string tag, string descricao, string alerta)
+		{
+			// O Formato da mensagem de erro pode ser alterado pelo usuario alterando-se a property FormatoAlerta: onde;
+			// %TAG%       : Representa a TAG; ex: <nLacre>
+			// %ID%        : Representa a ID da TAG; ex X34
+			// %MSG%       : Representa a mensagem de alerta
+			// %DESCRICAO% : Representa a Descrição da TAG
 
-        #endregion Propriedades
+			var s = FormatoAlerta.Clone() as string;
+			if (s == null)
+				return;
 
-        #region Methods
+			s = s.Replace("%ID%", id).Replace("%TAG%", $"<{tag}>").Replace("%DESCRICAO%", descricao).Replace("%MSG%", alerta);
 
-        internal void AddAlerta(string id, string tag, string descricao, string alerta)
-        {
-            // O Formato da mensagem de erro pode ser alterado pelo usuario alterando-se a property FormatoAlerta: onde;
-            // %TAG%       : Representa a TAG; ex: <nLacre>
-            // %ID%        : Representa a ID da TAG; ex X34
-            // %MSG%       : Representa a mensagem de alerta
-            // %DESCRICAO% : Representa a Descrição da TAG
+			ErrosAlertas.Add(s);
+		}
 
-            var s = FormatoAlerta.Clone() as string;
-            if (s == null)
-                return;
-
-            s = s.Replace("%ID%", id).Replace("%TAG%", $"<{tag}>").Replace("%DESCRICAO%", descricao).Replace("%MSG%", alerta);
-
-            ErrosAlertas.Add(s);
-        }
-
-        #endregion Methods
-    }
+		#endregion Methods
+	}
 }
