+++ src/ACBr.Net.DFe.Core.Tests/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="ACBr.Net.Core" version="1.0.1.9-beta1" targetFramework="net45" />
-  <package id="Microsoft.NETCore.Platforms" version="2.0.0" targetFramework="net45" />
+  <package id="Microsoft.NETCore.Platforms" version="2.0.1" targetFramework="net45" />
   <package id="NETStandard.Library" version="2.0.0" targetFramework="net45" />
   <package id="System.Collections" version="4.3.0" targetFramework="net45" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net45" />
@@ -32,12 +32,12 @@
   <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net45" />
   <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="net45" />
   <package id="System.Xml.XDocument" version="4.3.0" targetFramework="net45" />
-  <package id="xunit" version="2.3.0" targetFramework="net45" />
+  <package id="xunit" version="2.3.1" targetFramework="net45" />
   <package id="xunit.abstractions" version="2.0.1" targetFramework="net45" />
-  <package id="xunit.analyzers" version="0.7.0" targetFramework="net45" />
-  <package id="xunit.assert" version="2.3.0" targetFramework="net45" />
-  <package id="xunit.core" version="2.3.0" targetFramework="net45" />
-  <package id="xunit.extensibility.core" version="2.3.0" targetFramework="net45" />
-  <package id="xunit.extensibility.execution" version="2.3.0" targetFramework="net45" />
-  <package id="xunit.runner.visualstudio" version="2.3.0" targetFramework="net45" developmentDependency="true" />
+  <package id="xunit.analyzers" version="0.8.0" targetFramework="net45" />
+  <package id="xunit.assert" version="2.3.1" targetFramework="net45" />
+  <package id="xunit.core" version="2.3.1" targetFramework="net45" />
+  <package id="xunit.extensibility.core" version="2.3.1" targetFramework="net45" />
+  <package id="xunit.extensibility.execution" version="2.3.1" targetFramework="net45" />
+  <package id="xunit.runner.visualstudio" version="2.3.1" targetFramework="net45" developmentDependency="true" />
 </packages>
