+++ src/ACBr.Net.DFe.Core.sln
@@ -49,7 +49,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{D5613362-449B-4C9D-AC82-7E234CFE545A} = {BD191AE8-0354-4749-A543-3B0E28310332}
 		{2E3DF5C2-8A38-4A03-86D7-8D463C917E47} = {63B2512D-A2E1-4376-A984-B527D285242A}
-		{A391A2ED-C1F2-472C-93DA-0C493BA8EA6B} = {B0D440BA-99BE-47A2-9791-8B56D012B791}
 		{6EBE3041-2A58-4760-BCEE-A26C90413058} = {BD191AE8-0354-4749-A543-3B0E28310332}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
