+++ src/ACBr.Net.DFe.Core/packages.config
@@ -2,5 +2,5 @@
 <packages>
   <package id="ACBr.Net.Core" version="1.0.1.9-beta1" targetFramework="net40" />
   <package id="ExtraConstraints.Fody" version="1.11.0" targetFramework="net40" developmentDependency="true" />
-  <package id="Fody" version="2.2.0" targetFramework="net40" developmentDependency="true" />
+  <package id="Fody" version="2.2.1.0" targetFramework="net40" developmentDependency="true" />
 </packages>
